Jmol.___JmolVersion="15.2.17"   also 14.32.17
bug fix: Show $s2  for isosurface reports first isosurface only
bug fix: _M.unitCellParams does not update after modifying the unit cell
bug fix: GausssianReader Tensor "csa" broken
bug fix: ScriptEditor many issues
bug fix: ... HKL {1 1 1}  works, but HKL 1 1 1 fails, if not last in a
command

new feature: getproperty("atomInfo").select("fx,fy,fz")
 -- adds fractional coordinates to atomInfo

new feature: draw boundbox @1 @2 @3 @4...
 -- draws a boundbox with edges in the x, y, and z directions through a
set of points

new feature: draw best boundbox @1 @2 @3 @4...
 -- draws the best boundbox through a set of points

new feature: draw best plane @1 @2 @3 @4...
 -- draws the best plane through a set of points

new feature: draw best line @1 @2 @3 @4...
 -- draws the best line through a set of points

new feature: draw intersection boundbox [plane description]
 -- include PLANE and HKL

new feature: draw intersection [unitcell or boundbox description] [plane
description]
 -- intersects all types of boundboxes and unitcells with any type of
plane

new feature: draw intersection [unitcell or boundbox description] ON
[plane description]
 -- projects a unitcell or boundbox on any plane

new feature: draw intersection [unitcell or boundbox description] LINE
@1 @2
 -- extends the specified line to intersect a unitcell or boundbox

new feature: draw diameter -1 ....
  -- draws a simple line with 1-pixel width

new feature: draw diameter -n ....
  -- draws a dotted 1-pixel line with alternating n on, n off pixels

new feature: draw HKL {1 1 1 x.x}
  -- offsets an HKL plane from the origin by x.x Angstroms

new feature: draw HKL {1 1 1} OFFSET x.x
  -- alternative to HCL {1 1 1 x.x}

JmolVersion="15.2.16"
  also 14.32.16

new feature: (JavaScript API) Jmol.setCallback(jmolApplet0, "xxx",
function() {......})
 -- where "xxx" is one of the allowed callbacks, namely (case
insensitive):
    ANIMFRAME, APPLETREADY, ATOMMOVED, AUDIO, CLICK, DRAGDROP, ECHO,
    ERROR, EVAL, HOVER, IMAGE, LOADSTRUCT, MEASURE, MESSAGE,
MINIMIZATION,
    MODELKIT, PICK, RESIZE, SCRIPT, SELECT, SERVICE, STRUCTUREMODIFIED,
SYNC
 -- allows anonymous function setting for callbacks rather than using
JmolScript SET xxxCallback "funcName"

new feature: (JavaScript) Info = { .... , errorCallback:
function{....},...}

 -- allows setting of functions rather than function names in Info
parameters

new feature: Drag-Drop into Java AppConsole assumed to be script and run

new feature: Drag-Drop into JavaScript output console assumed to be
script and run

new feature: Drag-Drop into JavaScript input console displayed

+++ src/javajs/util/VideoReader.java
@@ -138,13 +138,13 @@ public class VideoReader {
 		case "smhd":
 		case "vmhd":
 			thisTrackMap.put("_trackType", (blockType.equals("smhd") ? "sound" : "video"));
-			// fall through;
-		default:
+      //$FALL-THROUGH$
+    default:
 			skip(len);
 			len = 0;
 			break;
 		}
-		return (int) len;
+		return len;
 	}
 
 	protected int readInner(int len, Map<String, Object> map) throws IOException {
